Fix name regex to exclude fname fname pattern
The regular expression had been "match the first name or the last name,
then match the lastname or the first name", thinking that that would hit
all cases.  But it also lets slip through the false positive with a
fname fname or lname lname.  For example, say the membership has a "John
Smith" in it, then a line that has "John John" or "Smith Smith" will be
matched.

Change the regex to match first then last OR last then first.

Addresses-Issue:  #8