Fix coolrunning false positives
Tighten up the regex a bit.  Instead of "firstname-or-lastname, OR
fistname-or-lastname" (which could result in firstname-firstname), use
instead "firstname-lastname OR lastname-firstname".

Allow just a single space between first and last names.  That way we
don't get false positive caused by a trailing word that just so happens
to correspond to a name.

Closes-Issue:  #8